fs: readdir: update file hiding code
Added an array of files that we want to hide, might get moved to kconfig in future.
Added logger to see which files were blocked.
Switched to strncmp instead of strstr to avoid hiding fine files.
Removed namelen parameter, we are not going to use it.
Added a preset of files to hide to render lspeed and fde.ai modules useless.

No measurable performance degradation has been observed.

Signed-off-by: Yaroslav Furman <yaro330@gmail.com>